Update css notes
We're going to rearrange the folders and a few files in the near future. So add a few notes about where natural breaks and groups are in the current file setup.

For instance, "Modules" are used to group functions and variables in programming, but CSS IS NOT A PROGRAMMING LANGUAGE!

So simple, sensible, folders and files to come.