A lot of changes under the hood
The CSS is more or less the same, but I've overhauled the file and folder names:

1. We're beginning our `GPS` with `ECSS` characteristics.
2. `base/html` and `base/variables`
3. `globals/` files (this is where your design system would be
4. `pages/` folder for `#page` elements and `#sections` within them
5. Rename `#1` numbers to `1.` like they are in Markdown
6. Pedantically add fancy header comments and a MIT license note, right at the top, in prime position ;)
7. Remove the `-markdown.less` specific styles. Keep everything in the same `print-first.css` file from now on.
8. Moved some of the `_reset.less` files into `_root.less` and `text-elements.less`.

Sticking with the plural `globals` and `pages` for now. Might change my mind! Will also need to hunt for a more up-to-date `normalize.less` setup as it's getting quite old and possibly outdated now.

But wherever possible, DON'T BOTHER WITH CSS as it's a fucking ballache. Use your time more wisely on fun, artistic, and purposeful things.