Bump to a new minor version
It _could_ be more of a major version, but not a huge amount has changed in the actual compiled file, so child themes should be safe to update.

If you're using the `.grid-baseline()` mixin, it's gone now. Just add the `.gl-BaselineGrid` class to whichever element you want a grid baseline on. Override the styles with `--baseline` variables _inside_ the baseline class.